tracklisting/requirements.txt
# Tracklisting bot dependencies
requests>=2.28.0
pydub>=0.25.1
ShazamAPI>=0.0.2
PyYAML>=6.0

# pydub requires ffmpeg to be installed and on PATH for MP3 decoding.
# Install from https://ffmpeg.org/download.html or via: winget install ffmpeg

# ---------------------------------------------------------------------------
# Legacy GUI application (main_old.py) — additional dependencies required:
#   PySide6>=6.2.2
#   PyAudio>=0.2.13
#   numpy>=1.21.3
#   dataclass-csv>=1.3.0
# ---------------------------------------------------------------------------
